# docs/openapi/aiops_public_v1.json
{
  "components": {
    "schemas": {
...
      "ArtifactRef": {
        "additionalProperties": false,
        "properties": {
          "artifact_id": {
            "format": "uuid",
            "title": "Artifact Id",
            "type": "string"
          },
          "artifact_type": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Artifact Type",
            "type": "string"
          },
          "content_hash": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Hash",
            "type": "string"
          },
          "schema_version": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "artifact_id",
          "artifact_type",
          "schema_version",
          "content_hash"
        ],
        "title": "ArtifactRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpsRunCreate": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "format": "uuid",
            "title": "Agent Id",
            "type": "string"
          },
          "client_metadata": {
            "additionalProperties": true,
            "title": "Client Metadata",
            "type": "object"
          },
          "input": {
            "maxLength": 32000,
            "minLength": 1,
            "title": "Input",
            "type": "string"
          },
          "observation_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation End"
          },
          "observation_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observation Start"
          },
          "schema_version": {
            "default": "aiops.public.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "session_id": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "target_id",
          "input"
        ],
        "title": "OpsRunCreate",
        "type": "object"
      },
      "OpsRunReceipt": {
        "additionalProperties": false,
        "properties": {
          "event_cursor": {
            "minimum": 0.0,
            "title": "Event Cursor",
            "type": "integer"
          },
          "events_url": {
            "title": "Events Url",
            "type": "string"
          },
          "ops_run_id": {
            "format": "uuid",
            "title": "Ops Run Id",
            "type": "string"
          },
          "row_version": {
            "minimum": 1.0,
            "title": "Row Version",
            "type": "integer"
          },
          "schema_version": {
            "default": "aiops.public.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OpsRunStatus"
          }
        },
        "required": [
          "ops_run_id",
          "status",
          "row_version",
          "event_cursor",
          "events_url"
        ],
        "title": "OpsRunReceipt",
        "type": "object"
      },
      "OpsRunResult": {
        "additionalProperties": false,
        "description": "对外返回 Run 的最终 Artifact 内容。",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "final_artifact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtifactRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "ops_run_id": {
            "format": "uuid",
            "title": "Ops Run Id",
            "type": "string"
          },
          "payload": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "root_cause_grade": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RootCauseGrade"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "default": "aiops.public.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/OpsRunStatus"
          }
        },
        "required": [
          "ops_run_id",
          "status"
        ],
        "title": "OpsRunResult",
        "type": "object"
      },
      "OpsRunStatus": {
        "enum": [
          "CREATED",
          "SCOPING",
          "OBSERVING",
          "DIAGNOSING",
          "WAITING_INPUT",
          "DIAGNOSED",
          "PROPOSING",
          "WAITING_APPROVAL",
          "EXECUTING",
          "VERIFYING",
          "ROLLBACK_PENDING",
          "COMPLETED",
          "DEGRADED",
          "REJECTED",
          "FAILED",
          "CANCELLED",
          "EXPIRED"
        ],
        "title": "OpsRunStatus",
        "type": "string"
      },
      "OpsRunSummary": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "format": "uuid",
            "title": "Agent Id",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "final_artifact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ArtifactRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "ops_run_id": {
            "format": "uuid",
            "title": "Ops Run Id",
            "type": "string"
          },
          "root_cause_grade": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RootCauseGrade"
              },
              {
                "type": "null"
              }
            ]
          },
          "row_version": {
            "minimum": 1.0,
            "title": "Row Version",
            "type": "integer"
          },
          "schema_version": {
            "default": "aiops.public.v1",
            "title": "Schema Version",
            "type": "string"
          },
          "source_proposal_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Proposal Id"
          },
          "source_result_artifact_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Result Artifact Id"
          },
          "status": {
            "$ref": "#/components/schemas/OpsRunStatus"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          },
          "trigger_type": {
            "$ref": "#/components/schemas/TriggerType"
          }
        },
        "required": [
          "ops_run_id",
          "agent_id",
          "target_id",
          "trigger_type",
          "status",
          "row_version",
          "created_at"
        ],
        "title": "OpsRunSummary",
        "type": "object"
      },
...
      "RootCauseGrade": {
        "enum": [
          "CONFIRMED",
          "PROBABLE",
          "POSSIBLE",
          "INCONCLUSIVE"
        ],
        "title": "RootCauseGrade",
        "type": "string"
      },
...
      "TriggerType": {
        "enum": [
          "CHAT",
          "API",
          "ROOT",
          "ALERT",
          "SCHEDULE"
        ],
        "title": "TriggerType",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "KBot AIOps Public Contract",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/apps/aiops/runs": {
      "post": {
        "operationId": "create_ops_run_api_v1_apps_aiops_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpsRunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsRunReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ops Run",
        "tags": [
          "AIOps"
        ]
      }
    },
    "/api/v1/apps/aiops/runs/{run_id}": {
      "get": {
        "operationId": "get_ops_run_api_v1_apps_aiops_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsRunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ops Run",
        "tags": [
          "AIOps"
        ]
      }
    },
    "/api/v1/apps/aiops/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_ops_run_api_v1_apps_aiops_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsRunReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Ops Run",
        "tags": [
          "AIOps"
        ]
      }
    },
    "/api/v1/apps/aiops/runs/{run_id}/events": {
      "get": {
        "operationId": "stream_ops_run_events_api_v1_apps_aiops_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Ops Run Events",
        "tags": [
          "AIOps"
        ]
      }
    },
...
    "/api/v1/apps/aiops/runs/{run_id}/result": {
      "get": {
        "operationId": "get_ops_run_result_api_v1_apps_aiops_runs__run_id__result_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpsRunResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ops Run Result",
        "tags": [
          "AIOps"
        ]
      }
    },
...
  }
}